Don't crash if the cookie pipe doesn't exist. Review URL: http://chromereview.prom.corp.google.com/1187035
git-svn-id: svn://chrome-svn/chromeos/trunk@233 06c00378-0e64-4dae-be16-12b19f9950a1

+++ test_cookie_pipe.py
@@ -83,7 +83,8 @@ def Authenticate(email, password):
   return cookies
 
 def WriteToPipe(pipe_path, data):
-  os.remove(pipe_path)
+  if os.path.exists(pipe_path):
+    os.remove(pipe_path)
   os.mkfifo(pipe_path)
   f = open(pipe_path, 'w')
   f.write(data)
